requirements.txt
Flask==2.3.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
gunicorn==21.2.*
pytest==7.4.*
requests==2.32.*
